feat(37456): Criação e atualização de usuário atualiza unidades vinculadas
Altera na API a criação e atualização de usuários para atualizar a lista
de unidades vinculadas ao usuário.

Feitas algumas refaturações de
código.
