+++ sme_ptrf_apps/users/tests/test_api_user/test_api_list_grupos_user.py
@@ -1,0 +1,31 @@
+import pytest
+
+pytestmark = pytest.mark.django_db
+
+def test_consulta_grupos(
+        jwt_authenticated_client_u2,
+        usuario_2,
+        visao_ue,
+        visao_dre,
+        visao_sme,
+        permissao1,
+        permissao2,
+        grupo_1,
+        grupo_2,
+        grupo_3):
+
+    response = jwt_authenticated_client_u2.get("/api/usuarios/grupos/?visao=DRE", content_type='application/json')
+    result = response.json()
+    esperado = [
+        {
+            "id": str(grupo_2.id),
+            "nome": grupo_2.name,
+            "descricao": grupo_2.descricao
+        },
+        {
+            "id": str(grupo_3.id),
+            "nome": grupo_3.name,
+            "descricao": grupo_3.descricao
+        }]
+
+    assert result == esperado
